[#391] Fetch questions on wrapper
