+++ frontend/src/pages/cases/store/case_visual.js
@@ -3,6 +3,7 @@ import { Store } from "pullstate";
 const defaultCaseVisualState = {
   questionGroups: [],
   totalIncomeQuestions: [],
+  incomeDataDrivers: [],
 };
 
 const CaseVisualState = new Store(defaultCaseVisualState);
