+++ frontend/src/pages/cases/visualizations/ChartIncomeGap.js
