+++ frontend/src/pages/cases/visualizations/CompareIncomeGap.js
