+++ frontend/src/pages/cases/visualizations/EnterIncomeDataVisual.js
