+++ frontend/src/pages/cases/visualizations/index.js
